requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
pyzmq>=25.0
